feat: 将VideoPostResource的visit更改为Long类型，VideoPost提供子视频的总播放量统计

+++ service-video/src/main/java/com/inkneko/nekowindow/video/entity/VideoPostResource.java
@@ -17,7 +17,7 @@ public class VideoPostResource {
     Long videoId;
     Long nkid;
     String title;
-    Integer visit;
+    Long visit;
     Integer state;
     String reviewFailedReason;
     LocalDateTime createdAt;
